requirements.txt
flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas
python-whois==0.8.0
dnspython==2.4.2
Werkzeug==3.0.1